refactor: split event structure module
