+++ backend/src/event/structure.js
@@ -1,170 +1,12 @@
-const { format } = require("./date");
-const eventId = require("./id");
-
-/**
- * Base class for deserialization errors
- */
-class TryDeserializeError extends Error {
-    /**
-     * @param {string} message - Human readable error message
-     * @param {string} field - The field that caused the error
-     * @param {unknown} value - The invalid value
-     * @param {string} [expectedType] - The expected type/format
-     */
-    constructor(message, field, value, expectedType) {
-        super(message);
-        this.name = "TryDeserializeError";
-        this.field = field;
-        this.value = value;
-        this.expectedType = expectedType;
-    }
-}
-
-/**
- * Error for missing required fields
- */
-class MissingFieldError extends TryDeserializeError {
-    /**
-     * @param {string} field - The missing field name
-     */
-    constructor(field) {
-        super(`Missing required field: ${field}`, field, undefined, "any");
-        this.name = "MissingFieldError";
-    }
-}
-
-/**
- * Error for invalid field types
- */
-class InvalidTypeError extends TryDeserializeError {
-    /**
-     * @param {string} field - The field with invalid type
-     * @param {unknown} value - The invalid value
-     * @param {string} expectedType - The expected type
-     */
-    constructor(field, value, expectedType) {
-        const actualType = Array.isArray(value) ? 'array' : typeof value;
-        super(
-            `Invalid type for field '${field}': expected ${expectedType}, got ${actualType}`,
-            field,
-            value,
-            expectedType
-        );
-        this.name = "InvalidTypeError";
-        this.actualType = actualType;
-    }
-}
-
-/**
- * Error for invalid field values
- */
-class InvalidValueError extends TryDeserializeError {
-    /**
-     * @param {string} field - The field with invalid value
-     * @param {unknown} value - The invalid value
-     * @param {string} reason - Why the value is invalid
-     */
-    constructor(field, value, reason) {
-        super(`Invalid value for field '${field}': ${reason}`, field, value, undefined);
-        this.name = "InvalidValueError";
-        this.reason = reason;
-    }
-}
-
-/**
- * Error for invalid object structure
- */
-class InvalidStructureError extends TryDeserializeError {
-    /**
-     * @param {string} message - Error message
-     * @param {unknown} value - The invalid structure
-     */
-    constructor(message, value) {
-        super(message, "root", value, "object");
-        this.name = "InvalidStructureError";
-    }
-}
-
-/**
- * Error for nested field validation failures
- */
-class NestedFieldError extends TryDeserializeError {
-    /**
-     * @param {string} parentField - The parent field containing the nested error
-     * @param {string} nestedField - The nested field that failed
-     * @param {unknown} value - The invalid value
-     * @param {string} reason - Why validation failed
-     */
-    constructor(parentField, nestedField, value, reason) {
-        super(
-            `Invalid nested field '${parentField}.${nestedField}': ${reason}`,
-            `${parentField}.${nestedField}`,
-            value,
-            undefined
-        );
-        this.name = "NestedFieldError";
-        this.parentField = parentField;
-        this.nestedField = nestedField;
-        this.reason = reason;
-    }
-}
-
-/**
- * @param {unknown} object
- * @returns {object is TryDeserializeError}
- */
-function isTryDeserializeError(object) {
-    return object instanceof TryDeserializeError;
-}
-
-/**
- * @param {unknown} object
- * @returns {object is MissingFieldError}
- */
-function isMissingFieldError(object) {
-    return object instanceof MissingFieldError;
-}
-
-/**
- * @param {unknown} object
- * @returns {object is InvalidTypeError}
- */
-function isInvalidTypeError(object) {
-    return object instanceof InvalidTypeError;
-}
-
-/**
- * @param {unknown} object
- * @returns {object is InvalidValueError}
- */
-function isInvalidValueError(object) {
-    return object instanceof InvalidValueError;
-}
-
-/**
- * @param {unknown} object
- * @returns {object is InvalidStructureError}
- */
-function isInvalidStructureError(object) {
-    return object instanceof InvalidStructureError;
-}
-
-/**
- * @param {unknown} object
- * @returns {object is NestedFieldError}
- */
-function isNestedFieldError(object) {
-    return object instanceof NestedFieldError;
-}
-
-/**
- * Checks if a value is a plain object with string keys.
- * @param {unknown} value
- * @returns {value is Record<string, unknown>}
- */
-function isStringRecord(value) {
-    return typeof value === "object" && value !== null && !Array.isArray(value);
-}
+const { serialize, deserialize, tryDeserialize } = require('./serialization');
+const {
+    isTryDeserializeError,
+    isMissingFieldError,
+    isInvalidTypeError,
+    isInvalidValueError,
+    isInvalidStructureError,
+    isNestedFieldError,
+} = require('./errors');
 
 /**
  * @typedef Modifiers
@@ -198,200 +40,6 @@ function isStringRecord(value) {
  * @property {string} description - A description of the event (required).
  * @property {Creator} creator - Who created the event.
  */
-
-/**
- * @param {Event} event - The event object to serialize.
- * @returns {SerializedEvent} - The serialized event object.
- */
-function serialize(event) {
-    // De-nominalize the event.
-    const date = format(event.date);
-    const id = event.id.identifier;
-    const realEvent = { ...event, date, id };
-    return realEvent;
-}
-
-/**
- * @param {SerializedEvent} serializedEvent - The serialized event object from JSON.
- * @returns {Event} - The deserialized event object.
- */
-function deserialize(serializedEvent) {
-    return {
-        ...serializedEvent,
-        id: eventId.deserialize(serializedEvent),
-        date: new Date(serializedEvent.date),
-        modifiers: serializedEvent.modifiers || {},
-    };
-}
-
-/**
- * Attempts to deserialize an unknown object into an Event.
- * Returns the Event on success, or a TryDeserializeError on failure.
- *
- * @param {unknown} obj - The object to attempt to deserialize
- * @returns {Event | TryDeserializeError} - The deserialized Event or error object
- */
-function tryDeserialize(obj) {
-    try {
-        // Basic type and property checks
-        if (!obj || typeof obj !== "object" || Array.isArray(obj)) {
-            return new InvalidStructureError(
-                "Object must be a non-null object and not an array",
-                obj
-            );
-        }
-
-        // Extract and validate each field individually
-        if (!("id" in obj)) return new MissingFieldError("id");
-        const id = obj.id;
-        if (typeof id !== "string") {
-            return new InvalidTypeError("id", id, "string");
-        }
-
-        if (!("date" in obj)) return new MissingFieldError("date");
-        const date = obj.date;
-        if (typeof date !== "string") {
-            return new InvalidTypeError("date", date, "string");
-        }
-
-        if (!("original" in obj)) return new MissingFieldError("original");
-        const original = obj.original;
-        if (typeof original !== "string") {
-            return new InvalidTypeError("original", original, "string");
-        }
-
-        if (!("input" in obj)) return new MissingFieldError("input");
-        const input = obj.input;
-        if (typeof input !== "string") {
-            return new InvalidTypeError("input", input, "string");
-        }
-
-        if (!("type" in obj)) return new MissingFieldError("type");
-        const type = obj.type;
-        if (typeof type !== "string") {
-            return new InvalidTypeError("type", type, "string");
-        }
-
-        if (!("description" in obj)) return new MissingFieldError("description");
-        const description = obj.description;
-        if (typeof description !== "string") {
-            return new InvalidTypeError("description", description, "string");
-        }
-
-        if (!("creator" in obj)) return new MissingFieldError("creator");
-        const creator = obj.creator;
-        if (!creator || typeof creator !== "object" || Array.isArray(creator)) {
-            return new InvalidTypeError("creator", creator, "object");
-        }
-
-        // Handle modifiers - defaults to {} if missing. When provided it must
-        // be an object and not an array. Falsy values like 0 should be
-        // considered invalid rather than treated as an empty object.
-        const hasModifiers = "modifiers" in obj;
-        /** @type {unknown} */
-        const rawModifiers = hasModifiers ? obj.modifiers : {};
-        if (
-            hasModifiers &&
-            (rawModifiers === null ||
-                typeof rawModifiers !== "object" ||
-                Array.isArray(rawModifiers))
-        ) {
-            return new InvalidTypeError("modifiers", rawModifiers, "object");
-        }
-        /** @type {Record<string, unknown>} */
-        let modifiers = {};
-        if (hasModifiers && isStringRecord(rawModifiers)) {
-            modifiers = rawModifiers;
-        }
-
-        // Manually validate and parse the date
-        const dateObj = new Date(date);
-        if (isNaN(dateObj.getTime())) {
-            return new InvalidValueError("date", date, "not a valid date string");
-        }
-
-        // Manually validate creator has required properties
-        if (!creator || typeof creator !== "object") {
-            return new InvalidTypeError("creator", creator, "object");
-        }
-        if (!("name" in creator)) {
-            return new NestedFieldError("creator", "name", creator, "missing required field");
-        }
-        if (!("uuid" in creator)) {
-            return new NestedFieldError("creator", "uuid", creator, "missing required field");
-        }
-        if (!("version" in creator)) {
-            return new NestedFieldError("creator", "version", creator, "missing required field");
-        }
-        
-        const creatorName = creator.name;
-        const creatorUuid = creator.uuid;
-        const creatorVersion = creator.version;
-        if (typeof creatorName !== "string") {
-            return new NestedFieldError("creator", "name", creatorName, "expected string");
-        }
-        if (typeof creatorUuid !== "string") {
-            return new NestedFieldError("creator", "uuid", creatorUuid, "expected string");
-        }
-        if (typeof creatorVersion !== "string") {
-            return new NestedFieldError("creator", "version", creatorVersion, "expected string");
-        }
-
-        // Manually validate modifiers
-        // Build modifiers object manually using Object.entries
-        const sourceEntries = Object.entries(modifiers);
-        const validatedEntries = [];
-        for (let i = 0; i < sourceEntries.length; i++) {
-            const entry = sourceEntries[i];
-            if (!entry || entry.length !== 2) {
-                return new InvalidValueError("modifiers", modifiers, "invalid entry structure");
-            }
-            const key = entry[0];
-            const value = entry[1];
-            if (typeof value !== "string") {
-                return new NestedFieldError("modifiers", key, value, "expected string value");
-            }
-            validatedEntries.push([key, value]);
-        }
-        const validatedModifiers = Object.fromEntries(validatedEntries);
-
-        // Create validated SerializedEvent object for eventId.deserialize
-        /** @type {SerializedEvent} */
-        const validatedSerializedEvent = {
-            id: id,
-            date: date,
-            original: original,
-            input: input,
-            type: type,
-            description: description,
-            creator: {
-                name: creatorName,
-                uuid: creatorUuid,
-                version: creatorVersion,
-            },
-            modifiers: validatedModifiers,
-        };
-        
-        const eventIdObj = eventId.deserialize(validatedSerializedEvent);
-        if (!eventIdObj || !eventIdObj.identifier) {
-            return new InvalidValueError("id", id, "failed to deserialize event ID");
-        }
-
-        // Create and return the Event object
-        return {
-            ...validatedSerializedEvent,
-            id: eventIdObj,
-            date: dateObj,
-        };
-    } catch (error) {
-        // Wrap any other errors (e.g., from eventId.deserialize) in InvalidValueError
-        return new InvalidValueError(
-            "unknown",
-            obj,
-            `Unexpected error during deserialization: ${error instanceof Error ? error.message : String(error)}`
-        );
-    }
-}
 
 module.exports = {
     serialize,
